+++ Plugins/Tweaks/Tweaks.hpp
@@ -23,6 +23,7 @@ class ItemChargesCost;
 class FixDispelEffectLevels;
 class AddPrestigeclassCasterLevels;
 class FixUnlimitedPotionsBug;
+class UnhardcodeShields;
 
 class Tweaks : public NWNXLib::Plugin
 {
@@ -49,6 +50,7 @@ private:
     std::unique_ptr<FixDispelEffectLevels> m_FixDispelEffectLevels;
     std::unique_ptr<AddPrestigeclassCasterLevels> m_AddPrestigeclassCasterLevels;
     std::unique_ptr<FixUnlimitedPotionsBug> m_FixUnlimitedPotionsBug;
+    std::unique_ptr<UnhardcodeShields> m_UnhardcodeShields;
 };
 
 }